Split report file and integrity modes
